chg: use xsl variable instead of document maps
Reading the stylesheet as an input document with document('') seems
to be sensitive to the processing context, e.g. it has to end up with
a valid URI that might not be available.

Replaced with a xsl:variable with the same information, since it
works the same but I think is an XSLT 2.0 feature, which is fine for
Saxon but not for MSXML which is stuck in XLST 1 land.